# contracts/openapi/client-api.yaml
openapi: 3.1.0
info:
  title: DOM Client API
  version: 1.0.0
  description: >
    Official client HTTP contract for Account/Auth/Session and Workspace/Project/Folder
    metadata lifecycle (docs 09 and 28). This document describes only the HTTP surface:
    routes, operationId, headers and error-response assembly. Payload shapes are
    referenced from canonical JSON Schemas under /contracts and are never restated
    inline; code generation is driven by contracts/registry.yaml, not paths.
servers:
  - url: https://api.dom.internal
    description: Production
components:
  responses:
    ErrorEnvelopeResponse:
      description: Canonical error envelope (doc 28 §29)
      content:
        application/json:
          schema:
            $ref: "../errors/error-envelope.schema.json"
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Stable key for safe retry of a logical mutating command (doc 28 §13)
      schema:
        type: string
        format: uuid
paths:
...
  /v1/workspaces:
    post:
      operationId: CreateWorkspace
      summary: Create Workspace and request initial Owner bootstrap (FR-WRL-001/011)
      tags: [workspace]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "../commands/workspace/create-workspace.schema.json"
      responses:
        "201":
          description: Workspace created with Permission-owned initial Owner bootstrap
          content:
            application/json:
              schema:
                $ref: "../commands/workspace/create-workspace.schema.json#/$defs/CreateWorkspaceResponse"
        "400":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "401":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "403":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "409":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "503":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
    get:
      operationId: ListWorkspaces
      summary: List Workspaces visible to the authenticated actor (FR-WRL-001)
      tags: [workspace]
      responses:
        "200":
          description: Visible Workspace metadata
          content:
            application/json:
              schema:
                $ref: "../queries/workspace/list-workspaces.schema.json"
        "401":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "503":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
  /v1/workspaces/{workspaceId}:
    parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          $ref: "../ids/ids.schema.json#/$defs/WorkspaceId"
    get:
      operationId: GetWorkspace
      summary: Get visible Workspace metadata (FR-WRL-001/002)
      tags: [workspace]
      responses:
        "200":
          description: Workspace metadata
          content:
            application/json:
              schema:
                $ref: "../queries/workspace/get-workspace.schema.json"
        "401":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "403":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "404":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
    patch:
      operationId: RenameWorkspace
      summary: Rename Workspace with canonical sibling comparison (FR-WRL-002)
      tags: [workspace]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "../commands/workspace/rename-workspace.schema.json"
      responses:
        "200":
          description: Workspace renamed
          content:
            application/json:
              schema:
                $ref: "../commands/workspace/rename-workspace.schema.json#/$defs/RenameWorkspaceResponse"
        "400":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "401":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "403":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "404":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "409":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "503":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
  /v1/workspaces/{workspaceId}/projects:
    parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          $ref: "../ids/ids.schema.json#/$defs/WorkspaceId"
    post:
      operationId: CreateProject
      summary: Create Project metadata (FR-WRL-003)
      tags: [workspace]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "../commands/workspace/create-project.schema.json"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "../commands/workspace/create-project.schema.json#/$defs/CreateProjectResponse"
        "400":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "401":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "403":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "404":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "409":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "503":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
    get:
      operationId: ListProjects
      summary: List Project metadata in a Workspace (FR-WRL-003/004)
      tags: [workspace]
      responses:
        "200":
          description: Project metadata list
          content:
            application/json:
              schema:
                $ref: "../queries/workspace/list-projects.schema.json"
        "401":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "403":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "404":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "503":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
  /v1/projects/{projectId}:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          $ref: "../ids/ids.schema.json#/$defs/ProjectId"
    get:
      operationId: GetProjectTree
      summary: Get metadata-only Project Tree (FR-WRL-004)
      tags: [workspace]
      responses:
        "200":
          description: Project and Folder metadata, without Resource content
          content:
            application/json:
              schema:
                $ref: "../queries/workspace/get-project-tree.schema.json"
        "401":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "403":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "404":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "503":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
    patch:
      operationId: RenameProject
      summary: Rename Project metadata (FR-WRL-003)
      tags: [workspace]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "../commands/workspace/rename-project.schema.json"
      responses:
        "200":
          description: Project renamed
          content:
            application/json:
              schema:
                $ref: "../commands/workspace/rename-project.schema.json#/$defs/RenameProjectResponse"
        "400":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "401":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "403":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "404":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "409":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "503":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
  /v1/projects/{projectId}/archive:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          $ref: "../ids/ids.schema.json#/$defs/ProjectId"
    post:
      operationId: ArchiveProject
      summary: Archive Project as read-only metadata (FR-WRL-007)
      tags: [workspace]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "../commands/workspace/archive-project.schema.json"
      responses:
        "200":
          description: Project archived
          content:
            application/json:
              schema:
                $ref: "../commands/workspace/archive-project.schema.json#/$defs/ArchiveProjectResponse"
        "401":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "403":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "404":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "409":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "503":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
  /v1/projects/{projectId}/unarchive:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          $ref: "../ids/ids.schema.json#/$defs/ProjectId"
    post:
      operationId: UnarchiveProject
      summary: Restore Project from Archive (FR-WRL-007)
      tags: [workspace]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "../commands/workspace/unarchive-project.schema.json"
      responses:
        "200":
          description: Project unarchived
          content:
            application/json:
              schema:
                $ref: "../commands/workspace/unarchive-project.schema.json#/$defs/UnarchiveProjectResponse"
        "401":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "403":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "404":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "409":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "503":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
  /v1/projects/{projectId}/trash:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          $ref: "../ids/ids.schema.json#/$defs/ProjectId"
    post:
      operationId: TrashProject
      summary: Trash Project subtree by ancestor lifecycle (FR-WRL-008)
      tags: [workspace]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "../commands/workspace/trash-project.schema.json"
      responses:
        "200":
          description: Project and descendants marked effectively Trashed by ancestor lifecycle
          content:
            application/json:
              schema:
                $ref: "../commands/workspace/trash-project.schema.json#/$defs/TrashProjectResponse"
        "401":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "403":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "404":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "409":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "503":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
  /v1/projects/{projectId}/restore:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          $ref: "../ids/ids.schema.json#/$defs/ProjectId"
    post:
      operationId: RestoreProject
      summary: Restore Project subtree by ancestor lifecycle (FR-WRL-008)
      tags: [workspace]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "../commands/workspace/restore-project.schema.json"
      responses:
        "200":
          description: Project and descendants restored to effective Active state
          content:
            application/json:
              schema:
                $ref: "../commands/workspace/restore-project.schema.json#/$defs/RestoreProjectResponse"
        "401":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "403":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "404":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "409":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "503":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
...
  /v1/folders/{folderId}:
    parameters:
      - name: folderId
        in: path
        required: true
        schema:
          $ref: "../ids/ids.schema.json#/$defs/FolderId"
    patch:
      operationId: RenameFolder
      summary: Rename Folder metadata (FR-WRL-005)
      tags: [workspace]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "../commands/workspace/rename-folder.schema.json"
      responses:
        "200":
          description: Folder renamed
          content:
            application/json:
              schema:
                $ref: "../commands/workspace/rename-folder.schema.json#/$defs/RenameFolderResponse"
        "400":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "401":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "403":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "404":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "409":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "503":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
...